requirements.txt
paho-mqtt
pyyaml
# Elasticsearch 2.x
elasticsearch>=2.0.0,<3.0.0
